+++ bindings/nodejs/examples/how_tos/nft_collection/00_mint_issuer_nft.ts
@@ -1,0 +1,68 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+import { MintNftParams, NftId, NftOutput, RegularTransactionEssence, TransactionPayload, utf8ToHex, Utils, Wallet } from '@iota/sdk';
+
+// In this example we will mint the issuer NFT for the NFT collection.
+//
+// Make sure that `STRONGHOLD_SNAPSHOT_PATH` and `WALLET_DB_PATH` already exist by
+// running the `how_tos/accounts-and-addresses/create-wallet` example!
+//
+// Rename `.env.example` to `.env` first, then run
+// yarn run-example ./how_tos/nfts/00_mint_issuer_nft.ts
+async function run() {
+    try {
+        if (!process.env.STRONGHOLD_PASSWORD) {
+            throw new Error(
+                '.env STRONGHOLD_PASSWORD is undefined, see .env.example',
+            );
+        }
+
+        // Create the wallet
+        const wallet = new Wallet({
+            storagePath: process.env.WALLET_DB_PATH,
+        });
+
+        // Get the account we generated with `01-create-wallet`
+        const account = await wallet.getAccount(
+            `${process.env.ACCOUNT_ALIAS_1}`,
+        );
+
+        await account.sync();
+        console.log(`Account synced!`);
+
+        // Issue the minting transaction and wait for its inclusion
+        console.log(`Sending NFT minting transaction...`);
+        const params: MintNftParams = {
+            immutableMetadata: utf8ToHex("This NFT will be the issuer from the awesome NFT collection"),
+        };
+        const prepared = await account.prepareMintNfts([params]);
+
+        const transaction = await prepared.send();
+
+        // Wait for transaction to get included
+        const blockId = await account.retryTransactionUntilIncluded(
+            transaction.transactionId,
+        );
+        console.log(
+            `Block included: ${process.env.EXPLORER_URL}/block/${blockId}`,
+        );
+
+        const essence: RegularTransactionEssence = (transaction.payload as TransactionPayload).essence;
+        essence.outputs.forEach(output: Output => {
+            if (output instanceof NftOutput) {
+                const nftOutput = output as NftOutput;
+                if nftOutput.getNnftId().isNull() {
+                    const outputId = await Utils.computeOutputId(transaction.transactionId, outputIndex);
+                    const nftId: NftId = outputId);
+                    console.log(`New minted NFT id: {nft_id}`);
+                }
+            }
+        });
+    } catch (error) {
+        console.log('Error: ', error);
+    }
+    process.exit(0);
+}
+
+run();
