+++ bindings/nodejs/examples/how_tos/nft_collection/01_mint_collection_nft.ts
